requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
bidict==0.23.1
certifi==2024.2.2
charset-normalizer==3.3.2
frozenlist==1.4.1
h11==0.14.0
idna==3.7
multidict==6.0.5
pydub==0.25.1
python-engineio==4.9.0
python-socketio==5.11.2
requests==2.31.0
simple-websocket==1.0.0
six==1.16.0
SpeechRecognition==3.10.4
typing_extensions==4.11.0
urllib3==2.2.1
wsproto==1.2.0
yarl==1.9.4
openai==1.26.0
pandas==2.0.3
numpy==1.25.2
scikit-learn==1.2.2
python-dotenv